+++ trunk/xorp/pim/pim_mrt.hh
@@ -12,7 +12,7 @@
 // notice is a summary of the XORP LICENSE file; the license in that file is
 // legally binding.
 
-// $XORP: xorp/pim/pim_mrt.hh,v 1.16 2005/05/27 20:37:03 pavlin Exp $
+// $XORP: xorp/pim/pim_mrt.hh,v 1.17 2005/08/18 15:38:47 bms Exp $
 
 
 #ifndef __PIM_PIM_MRT_HH__
@@ -203,9 +203,9 @@ public:
 				  const IPvX& pim_nbr_addr);
     void add_task_pim_nbr_gen_id_changed(uint32_t vif_index,
 					 const IPvX& pim_nbr_addr);
-    void add_task_assert_rpf_interface_wc(uint16_t old_rpf_interface_rp,
+    void add_task_assert_rpf_interface_wc(uint32_t old_rpf_interface_rp,
 					  const IPvX& group_addr);
-    void add_task_assert_rpf_interface_sg(uint16_t old_rpf_interface_s,
+    void add_task_assert_rpf_interface_sg(uint32_t old_rpf_interface_s,
 					  const IPvX& source_addr,
 					  const IPvX& group_addr);
     void add_task_receive_join_rp(uint32_t vif_index, const IPvX& rp_addr);
